pyproject.toml
[project]
name = "daphniadetectv2"
version = "0.1.0"
description = "A pipeline for scientific analysis of animals of the genus Daphnia with YOLO."
authors = [
    {name = "", email = ""},
]
dependencies = ["pandas>=2.2.3", "opencv-python>=4.8.0", "imutils>=0.5.4", "torch>=2.6.0", "torchvision>=0.21.0", "torchaudio>=2.6.0", "ultralytics>=8.4.14", "easyocr>=1.7.2", "numpy>=2.4.2", "pillow>=12.1.1", "sahi>=0.11.15", "roboflow>=1.2.16", "timm>=1.0.25"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "GPLv3"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
package-dir = {"" = "src"}
distribution = true
